feat(routing): retry with empty grid when path is not found
